requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.3
djangorestframework==3.15.2
eval_type_backport==0.2.0
filelock==3.16.1
frozenlist==1.5.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
numpy==2.1.3
pillow==10.4.0
propcache==0.2.0
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
sqlparse==0.5.1
tabulate==0.9.0
together==1.3.3
tqdm==4.67.0
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
